requirements.txt
aiosqlite==0.19.0
discord.py==2.2.2
python-dotenv==1.0.0
yt_dlp==2025.12.08
Requests==2.32.4
ffmpeg==1.4
PyNaCl==1.5.0
curl_cffi==0.13.0
audioop-lts==0.2.2